+++ tests/unit/test_examples.py
@@ -16,11 +16,7 @@ from pathlib import Path
 import pytest
 
 from swiss_ai_model_launch.cli.main import _build_parser, build_launch_args_from_advanced
-from swiss_ai_model_launch.launchers.framework import (
-    OCF_BOOTSTRAP_ADDR_DEV,
-    render_master,
-    render_rank_scripts,
-)
+from swiss_ai_model_launch.launchers.framework import render_master, render_rank_scripts
 
 _HAS_SHELLCHECK = shutil.which("shellcheck") is not None
 _REPO_ROOT = Path(__file__).resolve().parents[2]
@@ -67,8 +63,6 @@ def test_example_renders_valid_bash(tmp_path: Path, example_path: str):
         pytest.skip(f"Example not found: {example_path}")
 
     args = _parse_sml_advanced_script(full_path.read_text())
-    # Synthesise the launcher-supplied bits: in production these come from
-    # InitConfig + the FirecREST/Slurm launcher's account+partition.
     launch_args = build_launch_args_from_advanced(
         args,
         account="proj01-test",
@@ -116,44 +110,3 @@ def test_example_passes_shellcheck(tmp_path: Path, example_path: str):
             capture_output=True,
         )
         assert result.returncode == 0, f"shellcheck failed for {example_path} → {filename}:\n{result.stdout.decode()}"
-
-
-# ── OCF bootstrap addr CLI plumbing ──────────────────────────────────────
-
-
-def _minimal_advanced_args(*extra: str):
-    parser = _build_parser()
-    tokens = [
-        "advanced",
-        "--firecrest-system",
-        "clariden",
-        "--partition",
-        "normal",
-        "--serving-framework",
-        "sglang",
-        "--slurm-environment",
-        "/path/to/env.toml",
-        "--framework-args",
-        "--served-model-name vendor/model-abc",
-        *extra,
-    ]
-    return parser.parse_args(tokens)
-
-
-def test_advanced_default_leaves_bootstrap_addr_unset():
-    args = _minimal_advanced_args()
-    la = build_launch_args_from_advanced(args, account="proj01", partition="normal")
-    assert la.ocf_bootstrap_addr is None
-
-
-def test_advanced_dev_flag_selects_dev_bootstrap_addr():
-    args = _minimal_advanced_args("--dev")
-    la = build_launch_args_from_advanced(args, account="proj01", partition="normal")
-    assert la.ocf_bootstrap_addr == OCF_BOOTSTRAP_ADDR_DEV
-
-
-def test_advanced_explicit_addr_overrides_dev():
-    custom = "/ip4/10.0.0.42/tcp/43905/p2p/QmCustomPeerXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-    args = _minimal_advanced_args("--dev", "--otela-bootstrap-addr", custom)
-    la = build_launch_args_from_advanced(args, account="proj01", partition="normal")
-    assert la.ocf_bootstrap_addr == custom
